move minim interact endpoint into optimizer plugin
