+++ plugins/optimizer/coordinator/shared_schemas.py
@@ -17,18 +17,9 @@ from typing import Optional
 
 import marshmallow as ma
 import numpy as np
+from plugins.optimizer.shared.schemas import NumpyArray
 
 from qhana_plugin_runner.api.util import MaBaseSchema
-
-
-class NumpyArray(ma.fields.Field):
-    def _serialize(self, value, attr, obj, **kwargs):
-        if value is None:
-            return None
-        return {"data": value.tolist(), "dtype": str(value.dtype), "shape": value.shape}
-
-    def _deserialize(self, value, attr, data, **kwargs):
-        return np.array(value["data"], dtype=value["dtype"]).reshape(value["shape"])
 
 
 @dataclass
